Configure the file rotation settings through the YAML

+++ ddsrecorder_participants/src/cpp/recorder/mcap/McapHandler.cpp
@@ -231,9 +231,14 @@ void McapHandler::add_data(
         samples_buffer_size += sample.dataSize;
     }
 
-    if (mcap_file_size_ + samples_buffer_size > configuration_.max_file_size)
+    // std::cout << "Samples buffer size: " << samples_buffer_size << std::endl;
+    // std::cout << "Max file size: " << configuration_.mcap_output_settings.max_file_size << std::endl;
+
+    if (mcap_file_size_ + samples_buffer_size > configuration_.mcap_output_settings.max_file_size)
     {
         std::cout << "MAX FILE SIZE REACHED" << std::endl;
+
+        std::cout << "Mcap file size: " << mcap_file_size_ << std::endl;
 
         logInfo(DDSRECORDER_MCAP_HANDLER, "Max file size reached, closing file and opening a new one...");
         lock.unlock();
@@ -514,17 +519,25 @@ mcap::Timestamp McapHandler::now()
 
 void McapHandler::open_file_nts_()
 {
-    // Update the file id
-    mcap_file_id_ += 1;
-    mcap_file_id_ %= configuration_.max_files;
-
-    if (mcap_file_id_to_filename_.count(mcap_file_id_))
-    {
-        std::filesystem::remove(mcap_file_id_to_filename_[mcap_file_id_]);
-    }
-
     // Reset file size
     mcap_file_size_ = 0;
+
+    // Update the file id
+    mcap_file_id_ += 1;
+
+    // Rotate output files
+    if (configuration_.mcap_output_settings.file_rotation)
+    {
+        if (configuration_.mcap_output_settings.max_files > 0)
+        {
+            mcap_file_id_ %= configuration_.mcap_output_settings.max_files;
+        }
+
+        if (mcap_file_id_to_filename_.count(mcap_file_id_))
+        {
+            std::filesystem::remove(mcap_file_id_to_filename_[mcap_file_id_]);
+        }
+    }
 
     // Generate the filename
     mcap_filename_ = configuration_.mcap_output_settings.output_filepath + "/";
@@ -541,7 +554,7 @@ void McapHandler::open_file_nts_()
 
     mcap_filename_ += configuration_.mcap_output_settings.output_filename;
 
-    if (configuration_.max_files > 1)
+    if (configuration_.mcap_output_settings.max_files > 1)
     {
         // Include the file id in the filename
         mcap_filename_ += "~" + std::to_string(mcap_file_id_);
